refactor: chat-api 적용 완료
