+++ client/src/components/chat/Chat.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import SidePanel from './SidePanel/SidePanel';
 import MainPanel from './MainPanel/MainPanel';
 import { getUserInfo } from '../../api/user-api';
-import getChatRooms from '../../api/getChatRooms';
+import { getChatRooms } from '../../api/chat-api';
 import { io } from 'socket.io-client';
 
 const socket = io('http://34.64.44.34:3003/chat', {
